requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
python-dotenv
pypdf
pdfplumber
